Fix version replacement and refreshVersions task
Instead of replacing artifacts, potentially with wrong dependency
attributes (as was the case for platform dependencies for example),
the artifacts are now no longer removed, and just have their version
constraint edited.
Also, after dependency replacement, refreshVersions would see the
edited version, and wouldn't spot the managed dependencies versions,
but now, the versionPlaceholder is placed into the reject part of the
version constraint, so it can be detected by the refreshVersions
task. That prevents dependencies removal and unwanted upgrade that
would otherwise happen on running it and then running anything else
in the project.