✚ Add typesafety to dependencies.implementation()
This is a follow up on https://github.com/jmfayard/refreshVersions/pull/229

It focuses on typesafety only

There is an elegant Kotlin solution for every problem.

Since implentation is defined like this

```
fun DependencyHandler.`implementation`(dependencyNotation: Any): Dependency? =
    add("implementation", dependencyNotation)
```

Then we can overwrite this function for when the parameter is IsNotADependency or DependencyNotationAndGroup

This way the IDE shows you that made a mistake