✚ Add type safety to dependencies.implementation() (#237)
This is a follow up on https://github.com/jmfayard/refreshVersions/pull/229

It focuses on typesafety only

There is an elegant Kotlin solution for every problem.

The IDE now imports two version of `implementation`. 

```
fun DependencyHandler.`implementation`(dependencyNotation: Any): Dependency? =
    add("implementation", dependencyNotation)

@kotlin.Deprecated(ErrorIsNotADependency)
fun DependencyHandler.`implementation`(dependencyNotation: IsNotADependency): Dependency? =
    null
```

Since the second one is more specific, it takes precedence when the parameter is `IsNotADependency` or in practice `DependencyNotationAndGroup`

This way the IDE shows you that made a mistake

<img width="874" alt="GraySky____IdeaProjects_gray-sky-weather__-_build_gradle_kts___app_" src="https://user-images.githubusercontent.com/459464/91646121-1b9e6180-ea4c-11ea-87b1-c18cd08c63b1.png">