Fix implementation of ResettableDelegates.Lazy
It was missing the associatedDelegates.add(this) call.

This commit puts it in the parent class to avoid it being
forgotten for any future delegate that might be needed and
also remove code duplication/boilerplate.

This issue was causing the same OkHttpClient instance to be
reused across builds, and that was even more of an issue
since its executor was closed on build end, which made any
attempt to use it on the next build fail.

Fixes #263