Fix an edge case of the refreshVersions task
If a given version key wasn't already in the versions.properties file,
and the refreshVersions task was the first Gradle build to run,
the withDependencies callback would not be called before the
refreshVersions task would try to resolve the version, and it'd fail.

Now, it'll try to get it as when another Gradle build is run and the withDependencies callback is invoked.

This will get rid of the puzzling error messages like:
"Couldn't resolve version for ModuleId(group=whatever, name = something)"