Detect maven coordinates instead of matching any version
That ensures that we don't try to replace versions that might
be expected by various Gradle plugins or custom logic that
we won't handle anyway.

This commit also removes a hardcoded list of versions to
not replace, because we now play it safer and no longer need it.

This is currently limited to build.gradle(.kts) files.

Co-authored-by: Louis CAD <louis.cognault@gmail.com>