Fix and improve rejectVersionIf feature
Fixed the behavior on subsequent builds:
The RefreshVersionsConfigHolder instance can be reused across multiple
builds, even if the Gradle scripts changed. This caused the removal of
rejectVersionsIf to not be taken into account.
We fixed that by using the resettable delegates that we are already
using in other places in the RefreshVersionsConfigHolder.

This commit also moves down the version rejection notice header.

Co-authored-by: Louis CAD <louis.cognault@gmail.com>