Support plain http repositories
This will support the use case of local http maven repos.
Since we're not downloading the dependencies ourselves, but only checking
the versions, the worst that could happen under attack is bad data,
leading to a crash, hang, or out of memory error.
Downloading dependencies themselves over plain-text http requires opt-in
on Gradle side anyway.

Closes #456