Add support for Gradle configuration cache
This commit replaces calls to gradle.buildFinished with an AutoCloseable BuildService,
and marks the versions.properties file as used at configuration time,
so configuration is invalidated if its content changes, letting dependency updates
be applied if any.

However, this commit doesn't make refreshVersions own tasks compatible with
Gradle configuration cache, because it's almost impossible to do so.
Thankfully, this is not a big deal since that doesn't prevent the tasks from
completing, and having configuration cache enabled for the main refreshVersions
task wouldn't be helpful anyway since it modifies the versions.properties file,
which invalidates the configuration cache anyway.