Better support of versionFor
Usages of versionFor are now tracked,
to avoid marking an entry as unused in the
versions.properties file.
They are also taken into account when running
the refreshVersions task, and if the version entry
is missing, it's looked up just like for other
dependencies.

In other words, versionFor now works just like
other dependencies with the version placeholder.

Resolves #511