Update TomlUpdaterTest and its data
Now, we are testing with realistic data when it
comes to the versions retrieved from the repos,
and this will help ensure we are going in a code
path that does the proper filtering.

Also, we are now testing for the migration into
the version catalog of dependencies that don't
have a version specified (e.g. because specified
by a BoM).

Yes, the tests are currently failing.