Fix removal of rejectVersionIf not taken into account
The problem was that the RefreshVersionsExtension was
directly touching RefreshVersionsConfigHolder, whose instance is
reused by Gradle across runs.

Now, the rejectVersionIf predicate defaults to null, is taken
each time the plugin is applied to be set on the
RefreshVersionsConfigHolder, so if it's been removed, it's set back
to null as it should.