Fix findLineReferencing in VersionsCatalogUpdater
Before, it'd go for the first dependency, and if, for some reason,
it wasn't detected as used, it'd not appear in the dependenciesUpdates list,
which would prevent us from seeing available updates of actually used
dependencies that use the same version key.

This fixes it by looking for the first one where we have an update.