Fix UsedVersionForTracker for multi-project use
Before this fix, if versionFor was used in multiple modules
with different dependency notations, the last one would be the
only one retained, and in case the last one that happened to be called
was a version key instead of a dependency notation, it would simply
clear it, which means the updates for these used dependency notations
would all be filtered out.

This commit fixes that by using the rootProject as a reference.