Move the dependency notations webpage generating test
The "update the list of dependency notations on the website" is now
in the prePublishTest test suite, in the
DependencyNotationsWebpageUpdateTest.kt file.

This commit also renames the file from dependencies-notations.md to
dependency-notations.md. All references have been updated, except in
the README.md and the CHANGELOG.md files because we need to keep
pointing at working URLs until the new doc is deployed.

Finally, indent has been tweaked to produce more readable html.