Improve and fix version sorting
When we spot a non digit part that isn't recognized as
a stability marker, instead of treating it as unknown
stability and marking it less stable that everything else,
we now treat the dependency as stable, and the string is
considered to be a qualifier that takes part in the sorting.

This commit adds a bunch of test data that was helpful to
debug the changes from this commit.

Fixes #551
Closes #552